performance: optimize box filters

+++ src/common/box_filters.h
@@ -22,9 +22,9 @@
 #define BOX_ITERATIONS 8
 
 // ch = number of channels per pixel.  Supported values: 1 and 4
-void dt_box_mean(float *const buf, const int height, const int width, const int ch,
-                 const int radius, const int interations);
+void dt_box_mean(float *const buf, const size_t height, const size_t width, const int ch,
+                 const int radius, const unsigned interations);
 
-void dt_box_min(float *const buf, const int height, const int width, const int ch, const int radius);
-void dt_box_max(float *const buf, const int height, const int width, const int ch, const int radius);
+void dt_box_min(float *const buf, const size_t height, const size_t width, const int ch, const int radius);
+void dt_box_max(float *const buf, const size_t height, const size_t width, const int ch, const int radius);
